refactor: move state to SignInForm to remove prop drilling
